refactor: approval 개행 수정
